+++ pcs/common/str_tools.py
@@ -113,20 +113,22 @@ def format_name_value_id_list(
     return output
 
 
-def format_name_optional_value_list(
-    item_list: Sequence[tuple[str, Optional[str]]],
+def format_name_optional_value_or_id_list(
+    item_list: Sequence[tuple[str, Optional[str], Optional[str]]],
 ) -> list[str]:
     """
-    Turn 2-tuples to 'name=value' or 'name' strings with standard quoting
+    Turn 3-tuples to strings with starndard quoting:
+      * 'name'
+      * 'name=value'
+      * 'name (id: id)'
+      * 'name=value (id: id)'
     """
     output = []
-    for raw_name, raw_value in item_list:
+    for raw_name, raw_value, an_id in item_list:
         name = quote(raw_name, "= ")
-        if raw_value is None:
-            output.append(name)
-        else:
-            value = quote(raw_value, "= ")
-            output.append(f"{name}={value}")
+        value = f"={quote(raw_value, '= ')}" if raw_value is not None else ""
+        _id = f" (id: {an_id})" if an_id is not None else ""
+        output.append(f"{name}{value}{_id}")
     return output